chore: Update dependencies and fix minor issues
This commit updates several dependencies to their latest versions and includes minor fixes for string resources, calculations, and code cleanup.

Key changes:

-   **Dependency Updates (`gradle/libs.versions.toml`):**
    -   `composeHotReload`: `1.0.0-beta07` -> `1.0.0-beta09`
    -   `ksp`: `2.0.3` -> `2.0.2` (Note: this is a downgrade, likely for compatibility)
    -   `sqlite`: `2.6.0` -> `2.6.1`
    -   `room`: `2.8.0` -> `2.8.1`

-   **String Resource Fix (`values-de/strings.xml`):**
    -   Renamed `exercise_detail_content_description_default_equipment` to `exercise_detail_content_description_default_exercise` for better accuracy in the German localization.

-   **UI & ViewModel Fixes:**
    -   **`ExerciseDetailView.kt` & `WorkoutTemplateDetailView.kt`**: Updated views to use the corrected string resource `exercise_detail_content_description_default_exercise`.
    -   **`BeverageOverviewViewModel.kt`**: Corrected the logic for calculating `sumOf` and `avgAmount` for beverage entries by explicitly using `it.amount` instead of relying on a potentially incorrect `toInt()` or `toDouble()` conversion.

-   **Code Cleanup (`StartUpService.kt`):**
    -   Re-enabled the `seedDatabase(database)` call, which was previously commented out, to ensure the database is seeded on startup.

+++ composeApp/src/commonMain/kotlin/org/darthacheron/fitbe/health/beverages/BeverageOverviewViewModel.kt
@@ -133,7 +133,7 @@ class BeverageOverviewViewModel(
             .map { (date, beverageEntries) ->
                 BeverageOverview(
                     date,
-                    beverageEntries.sumOf { it.amount.toInt() }.toUInt()
+                    beverageEntries.sumOf { it.amount }.toUInt()
                 )
             }
     }
@@ -149,7 +149,7 @@ class BeverageOverviewViewModel(
             .mapNotNull { (_, group) ->
                 if (group.isEmpty()) return@mapNotNull null
 
-                val avgAmount = group.sumOf { it.amount.toDouble() } / group.size
+                val avgAmount = group.sumOf { it.amount } / group.size
 
                 BeverageOverview(
                     dateUtc = representativeDateSelector(group),
